Break into sevreral files and rename classes for clarity.
