+++ src/openstudio_lib/CMakeLists.txt
@@ -278,8 +278,10 @@ set(${target_name}_SRC
   ServiceWaterScene.hpp
   SimSettingsTabController.cpp
   SimSettingsTabController.hpp
-  GroundTemperatureWidget.cpp
-  GroundTemperatureWidget.hpp
+  GroundTemperatureMonthlyInspectorView.cpp
+  GroundTemperatureMonthlyInspectorView.hpp
+  GroundTemperatureView.cpp
+  GroundTemperatureView.hpp
   SimSettingsTabView.cpp
   SimSettingsTabView.hpp
   SimSettingsView.cpp
@@ -637,7 +639,8 @@ set(${target_name}_moc
   SimSettingsTabController.hpp
   SimSettingsTabView.hpp
   SimSettingsView.hpp
-  GroundTemperatureWidget.hpp
+  GroundTemperatureMonthlyInspectorView.hpp
+  GroundTemperatureView.hpp
   SpaceLoadInstancesWidget.hpp
   SpacesDaylightingGridView.hpp
   SpacesInteriorPartitionsGridView.hpp
