+++ src/openstudio_lib/GroundTemperatureMonthlyInspectorView.hpp
@@ -1,0 +1,74 @@
+/***********************************************************************************************************************
+*  OpenStudio(R), Copyright (c) OpenStudio Coalition and other contributors.
+*  See also https://openstudiocoalition.org/about/software_license/
+***********************************************************************************************************************/
+
+#ifndef OPENSTUDIO_GROUNDTEMPERATUREMONTHLYINSPECTORVIEW_HPP
+#define OPENSTUDIO_GROUNDTEMPERATUREMONTHLYINSPECTORVIEW_HPP
+
+#include <openstudio/model/ModelObject.hpp>
+#include <openstudio/model/SiteGroundTemperatureBuildingSurface.hpp>
+#include <openstudio/model/SiteGroundTemperatureShallow.hpp>
+
+#include <QWidget>
+
+#include <array>
+
+class QLabel;
+
+namespace openstudio {
+
+class OSQuantityEdit2;
+
+/** Abstract base for inspector widgets that show 12 monthly temperature fields. */
+class SiteGroundTemperatureMonthlyWidget : public QWidget
+{
+  Q_OBJECT
+
+ public:
+  explicit SiteGroundTemperatureMonthlyWidget(bool isIP, QWidget* parent = nullptr);
+  virtual ~SiteGroundTemperatureMonthlyWidget() = default;
+
+  virtual void attach(const model::ModelObject& obj) = 0;
+  void detach();
+
+ signals:
+  void toggleUnitsClicked(bool displayIP);
+
+ protected:
+  bool m_isIP;
+  QLabel* m_titleLabel = nullptr;
+  std::array<OSQuantityEdit2*, 12> m_edits{};
+};
+
+/** Inspector for Site:GroundTemperature:BuildingSurface — 12 monthly fields. */
+class SiteGroundTemperatureBuildingSurfaceWidget : public SiteGroundTemperatureMonthlyWidget
+{
+  Q_OBJECT
+
+ public:
+  explicit SiteGroundTemperatureBuildingSurfaceWidget(bool isIP, QWidget* parent = nullptr);
+
+  void attach(const model::ModelObject& obj) override;
+
+ private:
+  boost::optional<model::SiteGroundTemperatureBuildingSurface> m_obj;
+};
+
+/** Inspector for Site:GroundTemperature:Shallow — 12 monthly fields. */
+class SiteGroundTemperatureShallowWidget : public SiteGroundTemperatureMonthlyWidget
+{
+  Q_OBJECT
+
+ public:
+  explicit SiteGroundTemperatureShallowWidget(bool isIP, QWidget* parent = nullptr);
+
+  void attach(const model::ModelObject& obj) override;
+
+ private:
+  boost::optional<model::SiteGroundTemperatureShallow> m_obj;
+};
+
+}  // namespace openstudio
+
+#endif  // OPENSTUDIO_GROUNDTEMPERATUREMONTHLYINSPECTORVIEW_HPP
